more 1.9.2 agenda spotcheck wip

+++ src/main/java/gov/nysenate/openleg/client/view/agenda/AgendaIdView.java
@@ -5,7 +5,7 @@ import gov.nysenate.openleg.model.agenda.AgendaId;
 
 public class AgendaIdView implements ViewObject
 {
-    private int number;
+    private long number;
     private int year;
 
     public AgendaIdView(AgendaId agendaId) {
@@ -15,7 +15,7 @@ public class AgendaIdView implements ViewObject
         }
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
